Remove functionality for editing notes directly (Remove "_useLocalNoteContents")
After splitting the main big files in multipe modules, we have lots of
files referencing members of the main "plugin" object. One of those
modules is the amplenote_rw one, where we keep refering to
"this._useLocalNoteContents". I removed this field as well as all code
that edits amplenotes directly (maybe a multi-commit thing, if
everything is not resolved here). This is because we don't necessarily
want to edit notes directly anyway if it's more memory/time efficient to
do it in the in-memory cache (this._noteContents).